Added functionality for Editor tool selection
Added multiple modes for the editor which can
be selected from the gui via data-attributes
(data-editormode). Those attributes must just be
named the way the modes are named. The editor
automatically binds against the control element
and also sets or unsets an active class on the
selected mode. The editing mode for none (default),
add/edit district is already implemented.